feat: Conditionally render owner details in AnimalProfileDetails
Access `currentUser` context in `useGetAnimalProfiles`. Add currentUser to the dependency array of the useEffect hook, used for fetching data.

Update state to include an `isOwner` boolean value.

Create and set `initialState` as the default state for `animalProfileState`.

Add `isAdopted` field when creating an animal profile.

Conditionally render the owner details section in `AnimalProfileDetails` based on `isOwner` and `isAdopted`.