feat: Make health and owner info related form controlled
Use useForm to set initial healthAndOwnerInfo form values, and manage form data.

Pass useForm handlers as functions to the corresponding input events and use inputErrors to conditionally render errors.

Update nextHandler to check for errors on healthAndOwnerInfo form submit.

Create an onSubmitHandler. Pass onSubmitHandler to the on submit event of the healthAndOwnerInfo form.